Gather DeriveSha codes under derivesha package
