+++ blockchain/init_derive_sha.go
@@ -18,22 +18,22 @@ package blockchain
 
 import (
 	"github.com/klaytn/klaytn/blockchain/types"
-	"github.com/klaytn/klaytn/storage/statedb"
+	"github.com/klaytn/klaytn/blockchain/types/derivesha"
 )
 
 func InitDeriveSha(deriveShaImpl int) {
 	switch deriveShaImpl {
 	case types.ImplDeriveShaOriginal:
 		logger.Info("Using DeriveShaOrig!")
-		types.InitDeriveSha(statedb.DeriveShaOrig{})
+		types.InitDeriveSha(derivesha.DeriveShaOrig{})
 	case types.ImplDeriveShaSimple:
 		logger.Info("Using DeriveShaSimple!")
-		types.InitDeriveSha(types.DeriveShaSimple{})
+		types.InitDeriveSha(derivesha.DeriveShaSimple{})
 	case types.ImplDeriveShaConcat:
 		logger.Info("Using DeriveShaConcat!")
-		types.InitDeriveSha(types.DeriveShaConcat{})
+		types.InitDeriveSha(derivesha.DeriveShaConcat{})
 	default:
 		logger.Error("Undefined deriveShaImpl!! use DeriveShaOrig!")
-		types.InitDeriveSha(statedb.DeriveShaOrig{})
+		types.InitDeriveSha(derivesha.DeriveShaOrig{})
 	}
 }
