+++ blockchain/types/derive_sha.go
@@ -22,7 +22,6 @@ package types
 
 import (
 	"github.com/klaytn/klaytn/common"
-	"github.com/klaytn/klaytn/crypto/sha3"
 )
 
 type DerivableList interface {
@@ -51,61 +50,4 @@ func InitDeriveSha(i IDeriveSha) {
 
 func DeriveSha(list DerivableList) common.Hash {
 	return deriveShaObj.DeriveSha(list)
-}
-
-// An alternative implementation of DeriveSha()
-// This function generates a hash of `DerivableList` by simulating merkle tree generation
-type DeriveShaSimple struct{}
-
-func (d DeriveShaSimple) DeriveSha(list DerivableList) common.Hash {
-	hasher := sha3.NewKeccak256()
-
-	encoded := make([][]byte, 0, list.Len())
-	for i := 0; i < list.Len(); i++ {
-		hasher.Reset()
-		hasher.Write(list.GetRlp((i)))
-		encoded = append(encoded, hasher.Sum(nil))
-	}
-
-	for len(encoded) > 1 {
-		// make even numbers
-		if len(encoded)%2 == 1 {
-			encoded = append(encoded, encoded[len(encoded)-1])
-		}
-
-		for i := 0; i < len(encoded)/2; i++ {
-			hasher.Reset()
-			hasher.Write(encoded[2*i])
-			hasher.Write(encoded[2*i+1])
-
-			encoded[i] = hasher.Sum(nil)
-		}
-
-		encoded = encoded[0 : len(encoded)/2]
-	}
-
-	if len(encoded) == 0 {
-		hasher.Reset()
-		hasher.Write(nil)
-		return common.BytesToHash(hasher.Sum(nil))
-	}
-
-	return common.BytesToHash(encoded[0])
-}
-
-// An alternative implementation of DeriveSha()
-// This function generates a hash of `DerivableList` as below:
-// 1. make a byte slice by concatenating RLP-encoded items
-// 2. make a hash of the byte slice.
-type DeriveShaConcat struct{}
-
-func (d DeriveShaConcat) DeriveSha(list DerivableList) (hash common.Hash) {
-	hasher := sha3.NewKeccak256()
-
-	for i := 0; i < list.Len(); i++ {
-		hasher.Write(list.GetRlp(i))
-	}
-	hasher.Sum(hash[:0])
-
-	return hash
 }
