+++ blockchain/types/derivesha/simple.go
@@ -1,0 +1,47 @@
+package derivesha
+
+import (
+	"github.com/klaytn/klaytn/blockchain/types"
+	"github.com/klaytn/klaytn/common"
+	"github.com/klaytn/klaytn/crypto/sha3"
+)
+
+// An alternative implementation of DeriveSha()
+// This function generates a hash of `DerivableList` by simulating merkle tree generation
+type DeriveShaSimple struct{}
+
+func (d DeriveShaSimple) DeriveSha(list types.DerivableList) common.Hash {
+	hasher := sha3.NewKeccak256()
+
+	encoded := make([][]byte, 0, list.Len())
+	for i := 0; i < list.Len(); i++ {
+		hasher.Reset()
+		hasher.Write(list.GetRlp((i)))
+		encoded = append(encoded, hasher.Sum(nil))
+	}
+
+	for len(encoded) > 1 {
+		// make even numbers
+		if len(encoded)%2 == 1 {
+			encoded = append(encoded, encoded[len(encoded)-1])
+		}
+
+		for i := 0; i < len(encoded)/2; i++ {
+			hasher.Reset()
+			hasher.Write(encoded[2*i])
+			hasher.Write(encoded[2*i+1])
+
+			encoded[i] = hasher.Sum(nil)
+		}
+
+		encoded = encoded[0 : len(encoded)/2]
+	}
+
+	if len(encoded) == 0 {
+		hasher.Reset()
+		hasher.Write(nil)
+		return common.BytesToHash(hasher.Sum(nil))
+	}
+
+	return common.BytesToHash(encoded[0])
+}
